requirements.txt
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
pytest-ordering
allure-pytest
PyYAML
requests
openpyxl
PYHamcrest
pytest-json-report